[Libra Node] Move library logic out of main file.
Closes: #6143
